+++ components/Logo/FooterBrandLogo/FooterBrandLogo.component.vue
@@ -1,0 +1,14 @@
+<template lang="pug">
+a.footer-brand-logo(href="https://www.parolla.app" title="kelime oyunu" target="_blank") &copy; parolla
+</template>
+
+<script>
+import { defineComponent } from '@nuxtjs/composition-api'
+
+export default defineComponent({
+  props: {},
+  setup() {}
+})
+</script>
+
+<style lang="scss" src="./FooterBrandLogo.component.scss"></style>