+++ composables/useGameScene.js
@@ -647,7 +647,9 @@ export default () => {
       })
     }
 
-    carousels.alphabet.update()
+    if (activeGameMode.value !== gameModeKeyEnum.TOUR) {
+      carousels.alphabet.update()
+    }
   }
 
   const handleBeforeUnload = event => {