+++ pages/MusicMode/GuessTheSong/Play/index.vue
@@ -1,0 +1,15 @@
+<template lang="pug">
+.page.guess-the-song-play-page
+  GuessTheSongPlayScene
+</template>
+
+<script>
+import { defineComponent } from '@nuxtjs/composition-api'
+
+export default defineComponent({
+  layout: 'Default/Default.layout',
+  setup() {}
+})
+</script>
+
+<style lang="scss" src="./GuessTheSongPlay.page.scss"></style>